% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{as.character,DGVMUnit-method}
\alias{as.character,DGVMUnit-method}
\title{CF-conform unit string.}
\usage{
\S4method{as.character}{DGVMUnit}(x, ...)
}
\arguments{
\item{x}{An DGVMUnit object.}

\item{...}{Ignored further arguments.}
}
\value{
a character string
}
\description{
Returns a String from a DGVMUnit object, which can be used e.g. in ud.convert of package udunits2.
}
\examples{
x <- as.DGVMUnit("kW h m-2")
as.character(x)
units <- c("m2 s", "m2/s", "m^2*s", "C d")
x <- as.DGVMUnit(units)
sapply(x, as.character)
}
