% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/500Methods.R
\docType{methods}
\name{as.character,RVCUnit-method}
\alias{as.character,RVCUnit-method}
\title{CF-conform unit string.}
\usage{
\S4method{as.character}{RVCUnit}(x, ...)
}
\arguments{
\item{x}{An RVCUnit object.}

\item{...}{Ignored further arguments.}
}
\value{
a character string
}
\description{
Returns a String from a RVCUnit object, which can be used e.g. in ud.convert of package udunits2.
}
\examples{
x <- as.RVCUnit("kW h m-2")
as.character(x)
units <- c("m2 s", "m2/s", "m^2*s", "C d")
x <- as.RVCUnit(units)
sapply(x, as.character)
}

