\name{as.character}
\alias{as.character}
\title{CF-conform unit string.}
\description{
  Returns a String, which can be used ed in ud.convert of package udunits2.
}
\usage{
  as.character(x, ...)
}

\arguments{
  \item{x}{An RVCUnit object.}
 \item{...}{Ignored further arguments}
}
\value{
  A string
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
