% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.DGVMUnit}
\alias{as.DGVMUnit}
\title{Creates a DGVMUnit object from a unit string.}
\usage{
as.DGVMUnit(x = NA)
}
\arguments{
\item{x}{A string or a vector of strings parsable as units.}
}
\value{
An DGVMUnit object.
}
\description{
Creates a DGVMUnit object from a unit string.
}
\examples{
x <- as.DGVMUnit("kW h m-2")
units <- c("m2 s", "m2/s", "m^2*s", "C d")
x <- as.DGVMUnit(units)
}
