% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/500Methods.R
\docType{methods}
\name{as.expression,RVCUnit-method}
\alias{as.expression,RVCUnit-method}
\title{Expression for labels.}
\usage{
\S4method{as.expression}{RVCUnit}(x, ...)
}
\arguments{
\item{x}{An RVCUnit object.}

\item{...}{Ignored further arguments.}
}
\value{
an expression
}
\description{
Returns an expression from a RVCUnit object, which can be used e.g. as axis label.
}
\examples{
x <- as.RVCUnit("kW h m-2")
as.expression(x)
units <- c("m2 s", "m2/s", "m^2*s", "C d")
x <- as.RVCUnit(units)
sapply(x, as.expression)
}

