\name{as.expression}
\alias{as.expression}
\title{Unit for axis labels.}
\description{
  Returns a R expression, which can be used for axis labels
}
\usage{
  as.expression(x, ...)
}

\arguments{
  \item{x}{An RVCUnit object.}
  \item{...}{Ignored further arguments}
}
\value{
  A R expression
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
