% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200ClassFunctions.R
\name{as.RVCUnit}
\alias{as.RVCUnit}
\title{Creates a RVCUnit object from a unit string.}
\usage{
as.RVCUnit(x = NA)
}
\arguments{
\item{x}{A string or a vector of strings parsable as units.}
}
\value{
An RVCUnit object.
}
\description{
Creates a RVCUnit object from a unit string.
}
\examples{
x <- as.RVCUnit("kW h m-2")
units <- c("m2 s", "m2/s", "m^2*s", "C d")
x <- as.RVCUnit(units)
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

