% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cancel}
\alias{cancel}
\title{Cancel equal unit parts}
\usage{
cancel(x)
}
\arguments{
\item{x}{a \code{\linkS4class{DGVMUnit}} object}
}
\value{
a \code{\linkS4class{DGVMUnit}} object
}
\description{
Cancel equal unit parts
}
\examples{
x <- as.DGVMUnit("kg m s^-1 km^-2")
as.character(cancel(x))
}
