% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comparable}
\alias{comparable}
\title{Comarison two RVCUnit objects}
\usage{
comparable(a, b)
}
\arguments{
\item{a}{An RVCUnit object.}

\item{b}{An RVCUnit object.}
}
\value{
boolean
}
\description{
Checks if two RVCUnit objects are comparable (only different in unit prefixes).
}
\examples{
x <- as.RVCUnit("kg/m2")
y <- as.RVCUnit("kg ha^-2")
ret <- comparable(x,y)
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

