% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{divide}
\alias{divide}
\title{Division of two unit strings}
\usage{
divide(a, b)
}
\arguments{
\item{a}{An RVCUnit object.}

\item{b}{An RVCUnit object.}
}
\value{
A new RVCUnit object.
}
\description{
Divides two RVCUnit objects.
}
\examples{
x <- as.RVCUnit("kg/m2")
y <- as.RVCUnit("ha")
ret <- multiply(x,y)
as.character(ret)
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

