% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000ClassDef.R
\docType{class}
\name{RVCUnit-class}
\alias{RVCUnit-class}
\title{An S4 class holding the unit information.}
\description{
Each slot consists of a two element vector, holding the numeric unit prefix (multiplier) and exponent (the nominator is positive and the denominator negative).
}
\section{Slots}{

\describe{
\item{\code{g}}{for gram}

\item{\code{m}}{for meter}

\item{\code{W}}{for watts}

\item{\code{s}}{for seconds}

\item{\code{K}}{for Kelvin}
}}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

