\name{as.RVCUnit}
\alias{as.RVCUnit}
\title{Creates a RVCUnit object from a unit string.}
\description{
  Creates a RVCUnit object from a unit string.
}
\usage{
  as.RVCUnit(x=NA)
}

\arguments{
  \item{x}{A string or a vector of strings parsable as units.}
}
\value{
  An RVCUnit object.
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
