% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comparable}
\alias{comparable}
\title{Comarison two DGVMUnit objects}
\usage{
comparable(a, b)
}
\arguments{
\item{a}{An DGVMUnit object.}

\item{b}{An DGVMUnit object.}
}
\value{
boolean
}
\description{
Checks if two DGVMUnit objects are comparable (only different in unit prefixes).
}
\examples{
x <- as.DGVMUnit("kg/m2")
y <- as.DGVMUnit("kg ha^-2")
ret <- comparable(x,y)
}
