% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{divide}
\alias{divide}
\title{Division of two unit strings}
\usage{
divide(a, b)
}
\arguments{
\item{a}{An DGVMUnit object.}

\item{b}{An DGVMUnit object.}
}
\value{
A new DGVMUnit object.
}
\description{
Divides two DGVMUnit objects.
}
\examples{
x <- as.DGVMUnit("kg/m2")
y <- as.DGVMUnit("ha")
ret <- multiply(x,y)
as.character(ret)
}
