\name{divide}
\alias{divide}
\title{Division of two unit strings}
\description{
  Divides two RVCUnit objects.
}
\usage{
  divide(a, b)
}

\arguments{
  \item{a}{An RVCUnit object.}
  \item{b}{An RVCUnit object.}
}
\value{
  A new RVCUnit object.
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
