% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{equal}
\alias{equal}
\title{Comarison two DGVMUnit objects}
\usage{
equal(a, b)
}
\arguments{
\item{a}{An DGVMUnit object.}

\item{b}{An DGVMUnit object.}
}
\value{
boolean
}
\description{
Checks if two DGVMUnit objects are identical.
}
\examples{
x <- as.DGVMUnit("kg/m2")
y <- as.DGVMUnit("kg ha^-2")
ret <- equal(x,y)
}
