% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiply}
\alias{multiply}
\title{Multiplication of two RVCUnit objects}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{An RVCUnit object.}

\item{b}{An RVCUnit object.}
}
\value{
A new RVCUnit object.
}
\description{
Multiplies two RVCUnit objects.
}
\examples{
x <- as.RVCUnit("kg/m2")
y <- as.RVCUnit("kg m^-2 yr^-1")
ret <- divide(x,y)
as.character(ret)
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

