\name{multiply}
\alias{multiply}
\title{Multiplication of two RVCUnit objects}
\description{
  Multiplies two RVCUnit objects.
}
\usage{
  multiply(a, b)
}

\arguments{
  \item{a}{An RVCUnit object.}
  \item{b}{An RVCUnit object.}
}
\value{
  A new RVCUnit object.
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
